move tx add validator and change val weights
